+++ src/calva_mcp_server/extension/life_cycle_helpers.cljs
@@ -1,6 +1,6 @@
-(ns vsc-et.extension.life-cycle-helpers
+(ns calva-mcp-server.extension.life-cycle-helpers
   (:require ["vscode" :as vscode]
-            [vsc-et.extension.when-contexts :as when-contexts]))
+            [calva-mcp-server.extension.when-contexts :as when-contexts]))
 
 ;;;;; Extension lifecycle helper functions
 ;; These also assist with managing `vscode/Disposable`s in a hot-reloadable way.
@@ -15,7 +15,7 @@
   (swap! !state assoc :extension/disposables []))
 
 (defn cleanup! [!state]
-  (when-contexts/set-context!+ !state :vsc-et/active? false)
+  (when-contexts/set-context!+ !state :calva-mcp-server/active? false)
   (clear-disposables! !state))
 
 (defn register-command!