+++ src/calva_mcp_server/extension/when_contexts.cljs
@@ -1,4 +1,4 @@
-(ns vsc-et.extension.when-contexts
+(ns calva-mcp-server.extension.when-contexts
   (:require ["vscode" :as vscode]))
 
 (defn set-context!+ [!state k v]
@@ -9,5 +9,5 @@
   (get-in state [:extension/when-contexts k]))
 
 (comment
-  (require '[vsc-et.extension.db :as db])
-  (get-context db/!app-db :vsc-et/isActive))
+  (require '[calva-mcp-server.extension.db :as db])
+  (get-context db/!app-db :calva-mcp-server/active?))